Correct mocked capabilities

+++ src/main/java/dev/miku/r2dbc/mysql/constant/ServerStatuses.java
@@ -21,23 +21,23 @@ package dev.miku.r2dbc.mysql.constant;
  */
 public final class ServerStatuses {
 
-    public static final int IN_TRANSACTION = 1;
+    public static final short IN_TRANSACTION = 1;
 
-    public static final int AUTO_COMMIT = 2;
+    public static final short AUTO_COMMIT = 2;
 
-    public static final int MORE_RESULTS_EXISTS = 8;
+    public static final short MORE_RESULTS_EXISTS = 8;
 
-//    public static final int QUERY_NO_GOOD_INDEX_USED = 16;
-//    public static final int QUERY_NO_INDEX_USED = 32;
-//    public static final int CURSOR_EXISTS = 64;
-//    public static final int LAST_ROW_SENT = 128;
-//    public static final int DB_DROPPED = 256;
-//    public static final int NO_BACKSLASH_ESCAPES = 512;
-//    public static final int METADATA_CHANGED = 1024;
-//    public static final int QUERY_WAS_SLOW = 2048;
-//    public static final int PS_OUT_PARAMS = 4096;
-//    public static final int IN_TRANS_READONLY = 8192;
-//    public static final int SESSION_STATE_CHANGED = 16384;
+//    public static final short QUERY_NO_GOOD_INDEX_USED = 16;
+//    public static final short QUERY_NO_INDEX_USED = 32;
+//    public static final short CURSOR_EXISTS = 64;
+//    public static final short LAST_ROW_SENT = 128;
+//    public static final short DB_DROPPED = 256;
+//    public static final short NO_BACKSLASH_ESCAPES = 512;
+//    public static final short METADATA_CHANGED = 1024;
+//    public static final short QUERY_WAS_SLOW = 2048;
+//    public static final short PS_OUT_PARAMS = 4096;
+//    public static final short IN_TRANS_READONLY = 8192;
+//    public static final short SESSION_STATE_CHANGED = 16384;
 
     private ServerStatuses() {
     }
